requirements.txt
anyascii==0.3.1
asgiref==3.5.2
beautifulsoup4==4.8.2
bleach==5.0.1
bump2version==1.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.5
Django==3.1.14
django-cogwheels==0.3
django-filter==2.4.0
django-modelcluster==5.3
django-modeltranslation==0.18.2
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
docutils==0.18.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
keyring==23.6.0
l18n==2021.3
openpyxl==3.0.10
Pillow==9.1.1
pkginfo==1.8.3
polib==1.1.1
Pygments==2.12.0
pytz==2022.1
pywin32-ctypes==0.2.0
readme-renderer==35.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
tqdm==4.64.0
twine==3.8.0
typing-extensions==4.2.0
tzdata==2022.1
Unidecode==1.3.4
urllib3==1.26.9
wagtail==2.16.2
wagtail-localize==1.2.1
wagtail-modeltranslation==0.11.0
wagtailmenus==3.0.2
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==1.4.5
xlwt==1.3.0
zipp==3.8.0
